+++ application/core/Model_Entity_Controller_Helper.php
@@ -1,0 +1,11 @@
+<?php
+
+    /*
+        Interface for entities.
+        Functions to help controller with accessing and entities
+     */
+    interface Model_Entity_Controller_Helper {
+        public function getViewArray ();
+    }
+
+?>